update Progress tests to vitest
